requirements.txt
Flask
Flask-Session
requests
python-dotenv
pyheif
Pillow